[incomplete] image upload mostly working & project organization
